+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `\Mds\PimPrint\CoreBundle\InDesign\Command\CheckNewColumn` for dynamic columns breaks to be used in:
     - `\Mds\PimPrint\CoreBundle\InDesign\Command\SplitTable`
     - `\Mds\PimPrint\CoreBundle\InDesign\Command\GroupEnd`
+- Add `CheckNewPage` `marginOffset` for white space creation on dynamic layout breaks
+    - `\Mds\PimPrint\CoreBundle\InDesign\Command\CheckNewPage::setMarginOffset`
 
 ## [3.2.2] - 2024-01-16